experimental/Ambulance/ambulance-booking-app/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
djangorestframework==3.9.1
idna==2.8
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyJWT==1.7.1
pyotp==2.2.7
PySocks==1.6.8
pytz==2018.9
requests==2.21.0
six==1.12.0
twilio==6.24.1
urllib3==1.24.1
